renaming the module and candidate
