+++ examples/aggregate_signature/mod.rs
