+++ examples/aggregate_signature/signer.rs
@@ -1,5 +1,5 @@
-use crate::threshold_signature::registration::Registration;
-use crate::threshold_signature::signature::IndividualSignature;
+use crate::aggregate_signature::registration::Registration;
+use crate::aggregate_signature::signature::IndividualSignature;
 use blst::min_sig::{PublicKey, SecretKey};
 use rand_core::{CryptoRng, RngCore};
 use std::cmp::Ordering;
@@ -8,9 +8,9 @@ use std::cmp::Ordering;
 #[derive(Debug, Clone, Copy, Default)]
 pub struct VerificationKey(pub(crate) PublicKey);
 
-/// Threshold signature participant candidate
+/// Threshold signature candidate signer
 #[derive(Debug)]
-pub struct Candidate {
+pub struct Signer {
     /// Signing key of the type blst `SecretKey`
     signing_key: SecretKey,
     /// Signature verification key of the type blst `PublicKey`
@@ -20,7 +20,7 @@ pub struct Candidate {
 /// Registered Threshold signature signer.
 /// It includes signer's keypair and the registration check-sum.
 #[derive(Debug)]
-pub struct Signer {
+pub struct RegisteredSigner {
     signing_key: SecretKey,
     verification_key: VerificationKey,
     registration_check_sum: Vec<u8>,
@@ -68,7 +68,7 @@ impl Ord for VerificationKey {
     }
 }
 
-impl Candidate {
+impl Signer {
     /// Generate a new candidate, i.e., a bls signer
     pub fn new(rng: &mut (impl RngCore + CryptoRng)) -> Self {
         let mut ikm = [0u8; 32];
@@ -83,9 +83,9 @@ impl Candidate {
     }
 
     /// Create a new signer from candidate and closed registration
-    pub fn new_signer<const N: usize>(&self, registration: &Registration) -> Option<Signer> {
+    pub fn new_signer<const N: usize>(&self, registration: &Registration) -> Option<RegisteredSigner> {
         if registration.is_registered(&self.verification_key) {
-            Some(Signer {
+            Some(RegisteredSigner {
                 signing_key: self.signing_key.clone(),
                 verification_key: self.verification_key,
                 registration_check_sum: registration.check_sum.clone(),
@@ -96,7 +96,7 @@ impl Candidate {
     }
 }
 
-impl Signer {
+impl RegisteredSigner {
     /// Create a bls signature for given message.
     /// Signer creates a commitment by hashing the checksum of the closed registration and the message.
     /// The signature is generated by signing the commitment.
